Extract flow types to separate file
