+++ src/declarations.js
@@ -1,5 +1,5 @@
 /* @flow */
 
-import type { MuseGlobalType } from './types';
+import type { MuseGlobalType } from './lib/types';
 
 declare var __muse__ : MuseGlobalType;
