--- src/types.js
@@ -1,12 +1,0 @@
-/* @flow */
-
-export const TYPES = true;
-
-export type MuseServerConfigType = {|
-    assetsUrl : string
-|};
-
-export type MuseGlobalType = {|
-    serverConfig : MuseServerConfigType
-|};
-
